+++ vim/dotfiles/.config/nvim/lua/plugins/notify.lua
@@ -3,5 +3,6 @@ return {
     tag = "v3.13.2",
     config = function ()
         vim.notify = require("notify")
+        vim.notify.setup({background_colour = "#000000",})
     end
 }