+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.23] - 2024-01-09
+
+### Changed
+- [NVIM] - shell script executor: callback fn is injectable
+- [NVIM] - notifier: opts is injectable
+
 ## [0.0.22] - 2024-01-08
 
 ### Changed